made milestone index long->int

+++ src/main/java/com/iota/iri/model/Milestone.java
@@ -6,7 +6,7 @@ import java.util.Map;
  * Created by paul on 4/11/17.
  */
 public class Milestone {
-    public Long index;
+    public Integer index;
     public Hash hash;
     public Map<Hash, Long> snapshot;
 }
